Updated Order Processor functionality
Updated Orders.xml to have a faulty order
Updatd ScheduleImpl.java earliestAvailability method and
processingNumDays method
Updated Schedule.java to include processingNumDays method
Updated OrderProcessor functionality, created processOrder method with
recursion and print methods
Updated FacilityRecord to include processsingNumDays
Updated FacilityManager hasItem method
Updated FacilityImpl processingNumDays method
Updated Facility Interface processingNumDays method

+++ src/schedule/Schedule.java
@@ -32,6 +32,6 @@ public interface Schedule {
     int daysCapacity(int day);
     int earliestAvailability();
     int processingEndDate (int orderDay, int Qty);
-    int processingNumDays (int orderDay, int Qty);
+    double processingNumDays (int orderDay, int Qty);
     int reserve (int orderDay, int Qty);
 }
